requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cssselect2==0.8.0
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-unfold==0.56.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
eskiz-pkg==2.0.1
fonttools==4.57.0
idna==3.10
inflection==0.5.1
packaging==25.0
pillow==11.2.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
setuptools==80.3.1
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
wheel==0.45.1
zopfli==0.2.3.post1
